// src/app/models/utils.ts
export class Category{
    nid_category: number;
    sname_type: string;
    sdecription: string;
}

export class AmountsModel {
    nid_value: number;
    svalue: string;
    amount_income: number;
    amount_expense: number;
}

export class AmountMonthModel {
    ntype: number;
    snametype: string;
    namount: number;
    sname_type: string;
    nid_category: number;
}